python/requirements.txt
# Core LangChain and AI
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.20

# MCP (Model Context Protocol)
mcp>=0.1.0

# PDF Processing
pypdf2>=3.0.0
pdfplumber>=0.10.0
pymupdf>=1.23.0

# Manim and Manim Slides
manim>=0.18.0
manim-slides>=5.0.0

# Web Framework (for potential future use)
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
pillow>=10.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
